Add support for color images
Because it is dreadfully slow, a --color or --nocolor option will need to be added before this gets merged with the main branch.
A major cause of the slowdown is the sheer size of the palette, now that the number_of_colors changes the amount of shades per color channel.
Further optimizations will also have to be implemented eventually.